requirements.txt
fastapi
uvicorn
pydantic
Pillow
rembg
python-multipart
